% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_sheet_carry.R
\name{clean_sheet_cary}
\alias{clean_sheet_cary}
\title{Title}
\usage{
clean_sheet_cary(
  data = data,
  prefix = NULL,
  name_wave = "Wave",
  range_wave = c(380, 2500)
)
}
\arguments{
\item{data}{A data.frame containing the wavelength and reflectance values obtained from the Cary 5000 UV-Vis-IR spectrophotometers. It should be obtained directly from the export of a CSV file.}

\item{prefix}{Indicates the prefix defined for columns with no name.}

\item{name_wave}{Indicates the default name of the wavelength column.}

\item{range_wave}{Is the wavelength range used. The default is from 380 nm to 2500 nm.}
}
\value{
The function returns a data.frame with the columns organized, removing duplicate columns and placing the wavelength column in the first position of the data.frame.
}
\description{
Title
}
