% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cary.R
\docType{data}
\name{data_cary}
\alias{data_cary}
\title{Diffuse soil reflectance via Cary equipment}
\format{
A data frame with 5,417 rows and 52 columns:
\describe{
\item{\code{a1}, \code{a2}, \code{a3}, \code{a4}, \code{a5}, \code{a6}, \code{a7}, \code{a8}, \code{a9}, \code{a10}, \code{a11}, \code{a12}, \code{a13}, \code{a14}, \code{a15}, \code{a16}, \code{a17}, \code{a18}, \code{a19}, \code{a20}, \code{a21}, \code{a22}, \code{a23}, \code{a24}, \code{a26}}{Wavelength values in nanometers for each sample (labeled "a"). Each column represents the wavelength for a specific sample.}
\item{\code{x2}, \code{x4}, \code{x6}, \code{x8}, \code{x10}, \code{x12}, \code{x14}, \code{x16}, \code{x18}, \code{x20}, \code{x22}, \code{x24}, \code{x26}, \code{x28}, \code{x30}, \code{x32}, \code{x34}, \code{x36}, \code{x38}, \code{x40}, \code{x42}, \code{x44}, \code{x46}, \code{x48}, \code{x50}}{Reflectance percentage (\%R) corresponding to each wavelength for the same sample. The "x" columns represent the reflectance values for the corresponding "a" sample columns.}
}
}
\usage{
data_cary
}
\description{
A dataset containing diffuse reflectance measurements of soil samples from Brazil. The data were obtained using a Cary 5000 UV-Vis-NIR spectrophotometer, covering the spectral range from 2500 nm to 380 nm with a reading interval of 0.5 nm (Frosi et al., 2025).
}
\references{
Frosi, G., Inda, A., & Barrón, V. (2025).
Diffuse reflectance of soils from Brazil \emph{Data set}. Zenodo. \doi{10.5281/zenodo.14879041}
}
\keyword{datasets}
