% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relacion_hm_gt.R
\name{relation_hm_gt}
\alias{relation_hm_gt}
\title{Title}
\usage{
relation_hm_gt(
  data = data,
  smoothing_method = "cubic_spline",
  points_smoothing = 0.3,
  hem_go_limits = list(hem = c(535, 585), gt = c(430, 460)),
  peak_detect = "range",
  pv_tolerance = 10,
  name_wave = "wave",
  plot = TRUE,
  model_regression = "Vidal"
)
}
\arguments{
\item{model_regression}{}
}
\description{
Title
}
