% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_refle.R
\docType{data}
\name{soil_refle}
\alias{soil_refle}
\title{Tabela do índice SMP}
\format{
\subsection{\code{soil_refle}}{

Um data.frame com 4288 linha e 52 colunas:
\describe{
\item{Wavelength_nm}{comprimento de onda}
\item{xn}{amostra de solo}

...
}
}
}
\source{
"em busca"
}
\usage{
soil_refle
}
\description{
Dados do valor do índice SMP e a quantidade de calcário necessária para elevar o pH em 5,5, 6 e 6,5
}
\keyword{datasets}
